+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/robots/go2/go2_env_cfg.py
@@ -5,6 +5,7 @@ Copyright (c) 2024, Felipe Mohr Santos
 
 from omni.isaac.lab.utils import configclass
 
+from omni.isaac.lab_quadruped_tasks.robots import base_envs_cfg as base_envs
 from omni.isaac.lab_quadruped_tasks.cfg.quadruped_env_cfg import QuadrupedEnvCfg
 from omni.isaac.lab_quadruped_tasks.cfg.quadruped_terrains_cfg import (
     ROUGH_TERRAINS_CFG,
@@ -19,14 +20,13 @@ from omni.isaac.lab_assets.unitree import UNITREE_GO2_CFG
 
 import math
 
-
-######################
-# Go2 Base Environment
-######################
+#######################
+# Go2 Base Environments
+#######################
 
 
 @configclass
-class Go2BaseEnvCfg(QuadrupedEnvCfg):
+class Go2JointsBaseEnvCfg(base_envs.QuadrupedJointsEnvCfg):
     def __post_init__(self):
         super().__post_init__()
 
@@ -36,228 +36,108 @@ class Go2BaseEnvCfg(QuadrupedEnvCfg):
             ".*thigh_joint": math.pi / 4,
             ".*calf_joint": -math.pi / 2,
         }
-        self.scene.robot.init_state.pos = (0, 0, 0.32)
+
+        self.actions.action.scale = 0.2
+
+        self.rewards.rew_feet_air_time.weight = 0.75
+        self.rewards.pen_joint_powers.weight = -3e-3
+        self.rewards.pen_joint_deviation.weight = -0.1
+        self.rewards.pen_undesired_contacts.weight = -0.25
+
+
+@configclass
+class Go2CPGBaseEnvCfg(base_envs.QuadrupedCPGEnvCfg):
+    def __post_init__(self):
+        super().__post_init__()
+
+        self.scene.robot = UNITREE_GO2_CFG.replace(prim_path="{ENV_REGEX_NS}/Robot")
+        self.scene.robot.init_state.joint_pos = {
+            ".*hip_joint": 0.0,
+            ".*thigh_joint": math.pi / 4,
+            ".*calf_joint": -math.pi / 2,
+        }
+        # self.scene.robot.init_state.pos = (0, 0, 0.32)
         self.scene.robot.actuators["base_legs"].stiffness = 100.0
         self.scene.robot.actuators["base_legs"].damping = 2.0
 
-        self.rewards.pen_lin_vel_z.weight = -1.0
-        self.rewards.pen_ang_vel_xy.weight = -0.05
-
-        self.rewards.rew_lin_vel_xy.weight = 3.0
-        self.rewards.rew_ang_vel_z.weight = 1.5
-        self.rewards.pen_joint_powers.weight = -1e-3
-
-        self.rewards.rew_feet_air_time = None
-        self.rewards.pen_joint_deviation = None
-        self.rewards.pen_undesired_contacts = None
-        self.rewards.pen_feet_slide = None
-        self.rewards.pen_action_rate = None
-        self.rewards.pen_joint_accel = None
-        self.rewards.pen_flat_orientation = None
-        # self.terminations.base_contact = None
-
-        # self.actions.joint_pos.scale = 0.2
-
-        # self.rewards.rew_feet_air_time.weight = 0.75
-        # self.rewards.pen_joint_powers.weight = -3e-3
-        # self.rewards.pen_joint_deviation.weight = -0.1
-        # self.rewards.pen_undesired_contacts.weight = -0.25
-
-@configclass
-class Go2BaseEnvCfg_PLAY(Go2BaseEnvCfg):
-    def __post_init__(self):
-        super().__post_init__()
-
-        # make a smaller scene for play
-        self.scene.num_envs = 64
-
-        # disable randomization for play
-        self.observations.policy.enable_corruption = False
-        # remove random pushing event
-        self.events.push_robot = None
-        # remove random base mass addition event
-        self.events.add_base_mass = None
-
-
-############################
-# Go2 Blind Flat Environment
-############################
-
-
-@configclass
-class Go2BlindFlatEnvCfg(Go2BaseEnvCfg):
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.actions.cpg_action.use_joints_offset = False
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.events.change_vel_cmd = None
-        self.curriculum.terrain_levels = None
-
-
-@configclass
-class Go2BlindFlatEnvCfg_PLAY(Go2BaseEnvCfg_PLAY):
-    def __post_init__(self):
-        super().__post_init__()
-        self.actions.cpg_action.use_joints_offset = False
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.events.change_vel_cmd = None
-        self.curriculum.terrain_levels = None
-
-
-#############################
-# Go2 Blind Rough Environment
-#############################
-
-
-@configclass
-class Go2BlindRoughEnvCfg(Go2BaseEnvCfg):
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.events.change_vel_cmd = None
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.terrain_generator = ROUGH_TERRAINS_CFG
-
-        self.viewer.origin_type = "env"
-
-
-@configclass
-class Go2BlindRoughEnvCfg_PLAY(Go2BaseEnvCfg_PLAY):
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.events.change_vel_cmd = None
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.max_init_terrain_level = None
-        self.scene.terrain.terrain_generator = ROUGH_TERRAINS_PLAY_CFG
-
-
-##############################
-# Go2 Blind Stairs Environment
-##############################
-
-
-@configclass
-class Go2BlindStairsEnvCfg(Go2BaseEnvCfg):
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.commands.base_velocity.ranges.lin_vel_x = (0.5, 1.0)
-        self.commands.base_velocity.ranges.lin_vel_y = (0.0, 0.0)
-        self.commands.base_velocity.ranges.ang_vel_z = (-math.pi / 6, math.pi / 6)
-
-        self.events.reset_robot_base.params["pose_range"]["yaw"] = (0.0, 0.0)
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.terrain_generator = STAIRS_TERRAINS_CFG
-
-        self.viewer.origin_type = "env"
-
-
-@configclass
-class Go2BlindStairsEnvCfg_PLAY(Go2BaseEnvCfg_PLAY):
-    def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.height_scanner = None
-        self.observations.policy.height_map = None
-
-        self.commands.base_velocity.ranges.lin_vel_x = (0.5, 1.0)
-        self.commands.base_velocity.ranges.lin_vel_y = (0.0, 0.0)
-        self.commands.base_velocity.ranges.ang_vel_z = (0.0, 0.0)
-
-        self.events.reset_robot_base.params["pose_range"]["yaw"] = (0.0, 0.0)
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.max_init_terrain_level = None
-        self.scene.terrain.terrain_generator = STAIRS_TERRAINS_PLAY_CFG
+        # self.events.change_gait = None
+        # self.events.change_actuator_gains = None
 
 
 ########################
-# Go2 Vision Environment
+# Go2 Joints Environment
 ########################
 
 
 @configclass
-class Go2VisionEnvCfg(Go2BaseEnvCfg):
+class Go2JointsBlindFlatEnvCfg(Go2JointsBaseEnvCfg, base_envs.QuadrupedBlindFlatEnvCfg):
     def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.terrain_generator = FULL_TERRAINS_CFG
-
-        self.events.change_vel_cmd = None
-
-        self.viewer.origin_type = "env"
+        Go2JointsBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindFlatEnvCfg.__post_init__(self)
 
 
 @configclass
-class Go2VisionEnvCfg_PLAY(Go2BaseEnvCfg_PLAY):
+class Go2JointsBlindRoughEnvCfg(Go2JointsBaseEnvCfg, base_envs.QuadrupedBlindRoughEnvCfg):
     def __post_init__(self):
-        super().__post_init__()
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.max_init_terrain_level = None
-        self.scene.terrain.terrain_generator = FULL_TERRAINS_PLAY_CFG
-
-        self.events.change_vel_cmd = None
-
-        self.viewer.origin_type = "env"
-
-
-###############################
-# Go2 Vision Stairs Environment
-###############################
+        Go2JointsBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindRoughEnvCfg.__post_init__(self)
 
 
 @configclass
-class Go2VisionStairsEnvCfg(Go2BaseEnvCfg):
+class Go2JointsBlindStairsEnvCfg(Go2JointsBaseEnvCfg, base_envs.QuadrupedBlindStairsEnvCfg):
     def __post_init__(self):
-        super().__post_init__()
-
-        self.commands.base_velocity.ranges.lin_vel_x = (0.5, 1.0)
-        self.commands.base_velocity.ranges.lin_vel_y = (0.0, 0.0)
-        self.commands.base_velocity.ranges.ang_vel_z = (-math.pi / 6, math.pi / 6)
-
-        self.events.reset_robot_base.params["pose_range"]["yaw"] = (0.0, 0.0)
-
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.terrain_generator = STAIRS_TERRAINS_CFG
-
-        self.viewer.origin_type = "env"
+        Go2JointsBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindStairsEnvCfg.__post_init__(self)
 
 
 @configclass
-class Go2VisionStairsEnvCfg_PLAY(Go2BaseEnvCfg_PLAY):
+class Go2JointsVisionEnvCfg(Go2JointsBaseEnvCfg, base_envs.QuadrupedVisionEnvCfg):
     def __post_init__(self):
-        super().__post_init__()
+        Go2JointsBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedVisionEnvCfg.__post_init__(self)
 
-        self.commands.base_velocity.ranges.lin_vel_x = (0.5, 1.0)
-        self.commands.base_velocity.ranges.lin_vel_y = (0.0, 0.0)
-        self.commands.base_velocity.ranges.ang_vel_z = (0.0, 0.0)
 
-        self.events.reset_robot_base.params["pose_range"]["yaw"] = (0.0, 0.0)
+class Go2JointsVisionStairsEnvCfg(Go2JointsBaseEnvCfg, base_envs.QuadrupedVisionStairsEnvCfg):
+    def __post_init__(self):
+        Go2JointsBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedVisionStairsEnvCfg.__post_init__(self)
 
-        self.scene.terrain.terrain_type = "generator"
-        self.scene.terrain.max_init_terrain_level = None
-        self.scene.terrain.terrain_generator = STAIRS_TERRAINS_PLAY_CFG
 
-        self.viewer.origin_type = "env"
+#####################
+# Go2 CPG Environment
+#####################
+
+
+@configclass
+class Go2CPGBlindFlatEnvCfg(Go2CPGBaseEnvCfg, base_envs.QuadrupedBlindFlatEnvCfg):
+    def __post_init__(self):
+        Go2CPGBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindFlatEnvCfg.__post_init__(self)
+        self.actions.action.use_joints_offset = False
+
+
+@configclass
+class Go2CPGBlindRoughEnvCfg(Go2CPGBaseEnvCfg, base_envs.QuadrupedBlindRoughEnvCfg):
+    def __post_init__(self):
+        Go2CPGBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindRoughEnvCfg.__post_init__(self)
+
+
+@configclass
+class Go2CPGBlindStairsEnvCfg(Go2CPGBaseEnvCfg, base_envs.QuadrupedBlindStairsEnvCfg):
+    def __post_init__(self):
+        Go2CPGBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedBlindStairsEnvCfg.__post_init__(self)
+
+
+@configclass
+class Go2CPGVisionEnvCfg(Go2CPGBaseEnvCfg, base_envs.QuadrupedVisionEnvCfg):
+    def __post_init__(self):
+        Go2CPGBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedVisionEnvCfg.__post_init__(self)
+
+
+class Go2CPGVisionStairsEnvCfg(Go2CPGBaseEnvCfg, base_envs.QuadrupedVisionStairsEnvCfg):
+    def __post_init__(self):
+        Go2CPGBaseEnvCfg.__post_init__(self)
+        base_envs.QuadrupedVisionStairsEnvCfg.__post_init__(self)