+++ scripts/rsl_rl/train.py
@@ -12,8 +12,8 @@ import cli_args  # isort: skip
 # add argparse arguments
 parser = argparse.ArgumentParser(description="Train an RL agent with RSL-RL.")
 parser.add_argument("--video", action="store_true", default=False, help="Record videos during training.")
-parser.add_argument("--video_length", type=int, default=400, help="Length of the recorded video (in steps).")
-parser.add_argument("--video_interval", type=int, default=24000, help="Interval between video recordings (in steps).")
+parser.add_argument("--video_length", type=int, default=480, help="Length of the recorded video (in steps).")
+parser.add_argument("--video_interval", type=int, default=12000, help="Interval between video recordings (in steps).")
 parser.add_argument("--num_envs", type=int, default=None, help="Number of environments to simulate.")
 parser.add_argument("--max_iterations", type=int, default=None, help="Maximum number of iterations to train.")
 parser.add_argument("--save_interval", type=int, default=None, help="The number of iterations between saves")